Added 5-1 and 5-2.
Note: The .dsm files starts with Mario entering the level from the map.  To add this/view this:
A) Mario should already be standing on the spot of the level (you can start a movie from sram and move to that spot then copy paste this at the end)
B)enter a level by pressing "A", then stop the movie at the next frame, go to the .dsm file, replace the A press with the dsm file content.